Fixed to allow multiple actors of the same class.  Added Dining Philosophers Problem example app.
Previously instantiating multiple actors of the same class would violate an assertion.  The assertion is wrong and was removed.  The Framework was modified to support adding multiple actors of the same class.  When doing so, the post() method has to be modified to allow us to post messages to specific actors.

The Dining Philosophers Problem is a classic resource-contention problem.  I duplicated QP's solution to the problem which exposed the problem of allowing multiple actors of the same class.

Fixes #2