Fixed hang when running on system with high uptime
Problem is that when running pq on a computer with a high uptime, the software can hang (no response to Ctrl+C; must terminate the process).

Root cause is Framework.py:117 where sys.float_info.epsilon is added to the expiration time. When the system time is sufficiently large, adding epsilon will have no effect and the while loop will never exit (infinite loop).

Fix was to change the math so epsilon was added to the float's mantissa.  There is still room for improvement, but this solves the bug for now.

Fixes #3