Fixed VcdSpy, made Hsm.top a state
VcdSpy had a bug where the Spy statements were around Framework.run()'s call
to act.dispatch().  The problem was that act.state would be one thing on entry
and another on exit.  This left the VCD waveforms in an always-on state.
The fix was to put the Spy statements deeper into Hsm.dispatch so we can have a
pre and a post Spy handler.  This also makes it so that all involved states'
waveforms are turned on to the depth of the true handler.

Hsm.top had to have its pq_state attribute set.  Without it, an exception
could be thrown when the program was exited by ctrl+C.