Eliminated @staticmethod and "me" (BREAKS API)
Previously the Hsm class used @staticmethod with "me"
as the first argument, much like Python uses self.
The reason for this was to be more like old QP source code.
QP has since changed and this project has diverged enough from
QP to warrant a change to a more pythonic way of doing things.

This change breaks the API and apps will have to update.
The update is straightforward.  See this commits changes to
examples/*.py for how an app needs to change.

Also removed the old __init__simple.py.
It's in git history if you need it.