Fix a bug causing asserts / excess TimeEvents
There was some bad logic in _insort_time_event()
that allowed a TimeEvent with an expiration in the past
to cause two events to be issued.  This may result in the
assert "A TimeEvent must not be armed more than once."

The fix is to add an else, so the wrong logic doesn't
schedule a call_at() by mistake.

This likely fixes many instances where a TimeEvent with a period
that is too fast for the given load on the CPU
resulted in more than one event being dispatched.