Refactor internal data structures for handling TimeEvents
Previously a dict was used to hold the collection of active TimeEvents.
This was cumbersome.  It wasn't sorted and it prevented >1 TimeEvents
to be scheduled at the same time.

Now, two Lists are used.  One list holds the TimeEvents.  The other
list holds the next expiration time for each TimeEvent in the first list.
Both lists remain sorted by expiration time.  This means the first
entry in each list is the TimeEvent that will expire next.