Add a way to stop a specific AHSM
Currently, AHSMs can be added to the Framework
and they execute so long as the event loop runs.
This commit creates Ahsm.end() so an Ahsm can end
its own execution and remove itself from the Framework.
Calling .end() on an AHSM more than once without
calling .start() in-between will result in an exception.